recursion tests passed.
The permutations and powerset exercises need hints like split-into-monotonics as
they can be very discouraging in their trickiness. At the very least, remind the
student that apply exists (last spotted in pokerface), and provide a few more
warm ups with tree structure recursion. Maybe non-linear recursion should be
expanded for this and moved down. Or perhaps placed after sudoku, another set of
exercises similar to permutations in brute force searching.

Also, the encore should not be here. If you want a student to move along and
don't dwell and to come back later, then put it in its own exercise set for
later, then link it here for those who want to do it now. If the goal is
education, this is a fail. SICP did this better.